Phase 2: Update API, functionality and routes.
